phase-5/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for Todo application with JWT authentication and Neon PostgreSQL"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "pyjwt[crypto]>=2.9.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.2.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.3.0",
    "slowapi>=0.1.9",
    "python-json-logger>=2.0.7",
    "cachetools>=5.5.0",
    "psycopg2-binary>=2.9.11",
    "asyncpg>=0.31.0",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.9.0",
    "reportlab>=4.0.0",
    "openai-agents>=0.2.9",
    "mcp>=1.0.0",
    "openai>=1.0.0",
    "sse-starlette>=2.0.0",
    "nest-asyncio>=1.6.0",
    "openai-chatkit>=0.1.0",
    "dotenv>=0.9.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "black>=24.10.0",
    "ruff>=0.7.0",
]

[tool.black]
line-length = 100
target-version = ['py313']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | alembic
)/
'''

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "-v --strict-markers --cov=. --cov-report=term-missing"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pypdf2>=3.0.0",
]
